Imported source tree of CFS version 4.0.0.
At the time of importing, CFS was at v4.0.2. However source code is available back to v4.0.0 and this enables all changes from v4.0.0 to be recorded in the repository. This allows the tags branch to maintain accurate copies of all release 4 source trees.

Details of changes up to v4.0.2 are not recorded in this log - they can be found in other documentation.