Renamed CFS.phf as PreSVNProjectChanges.txt and reformatted information about which files changed in each version of the project.
Added new PreSVNFileChanges.txt file to record details of each change to each known file for which the information exists.

These files are for historical purposes only and maintain this information only as far as v4.0.2. Following r7 log messages in this repository along with the release files in the tags branch and the change log provide the same information.